Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 3653-xor-after-range-multiplication-queries-i/3653-xor-after-range-multiplication-queries-i.java
@@ -2,7 +2,7 @@ class Solution {
     public int xorAfterQueries(int[] nums, int[][] q) {
         int q_size = q.length;
         int ans = 0;
-        int mod = 1000000007;
+        long mod = 1000000007;
         for(int i=0; i<q_size; i++){
             int l= q[i][0];
             int r = q[i][1];
